move gettid to where it belongs
