+++ src/radical/utils/__init__.py
@@ -13,7 +13,7 @@ from .singleton      import Singleton
 from .process        import Process
 from .threads        import Thread, RLock
 from .threads        import is_main_thread, is_this_thread, cancel_main_thread
-from .threads        import main_thread, this_thread, get_thread_name
+from .threads        import main_thread, this_thread, get_thread_name, gettid
 from .threads        import set_cancellation_handler, unset_cancellation_handler
 from .threads        import raise_in_thread, ThreadExit, SignalRaised
 from .futures        import Future
