+++ src/radical/utils/threads.py
@@ -768,6 +768,24 @@ def get_thread_id():
 
 # ------------------------------------------------------------------------------
 #
+def gettid():
+    """
+    Python is not able to give us the native thread ID.  We thus use a syscall
+    to do so.  Since this is not portable, we fall back to None in case of error
+    (Hi MacOS).
+    """
+    try:
+        import ctypes
+        SYS_gettid = 186
+        libc = ctypes.cdll.LoadLibrary('libc.so.6')
+        return int(libc.syscall(SYS_gettid))
+    except:
+        return None
+
+
+
+# ------------------------------------------------------------------------------
+#
 def is_main_thread(t=None):
 
     if t:
